refactor: update architecture according to patterns and best practice
[1] https://dev-academy.com/angular-architecture-best-practices/

Close #22
