--- e2e/src/balance-form/balance-form.po.ts
@@ -1,27 +1,0 @@
-import { browser, by, element } from 'protractor';
-
-export class BalanceFormPage {
-  navigateTo() {
-    return browser.get('/refill/1');
-  }
-
-  form() {
-    return element(by.css('form'));
-  }
-
-  phoneNumber() {
-    return element(by.css('#phoneNumber'));
-  }
-
-  amount() {
-    return element(by.css('#amount'));
-  }
-
-  submit() {
-    return element(by.buttonText('Submit'));
-  }
-
-  message() {
-    return element(by.tagName('mat-hint'));
-  }
-}
