--- src/app/balance-form/balance-form.component.scss
@@ -1,5 +1,0 @@
-@import 'variables.scss';
-
-.warn {
-  color: mat-color($warnPalette, 300);
-}
