--- src/app/balance-form/index.ts
@@ -1,2 +1,0 @@
-export * from './balance-form-routing.module';
-export * from './balance-form.component';
