--- src/app/core/interfaces/index.ts
@@ -1,2 +1,0 @@
-export * from './provider';
-export * from './refill';
