--- src/app/core/interfaces/refill.ts
@@ -1,5 +1,0 @@
-export interface Refill {
-  providerId: string;
-  phoneNumber: string;
-  amount: number;
-}
