+++ src/app/home/components/provider-card/provider-card.component.scss
