+++ src/app/home/containers/index.ts
@@ -1,0 +1,1 @@
+export * from './home/home.component';
