--- src/app/home/index.ts
@@ -1,2 +1,0 @@
-export * from './home.component';
-export * from './home-routing.module';
