--- src/app/home/provider-card/index.ts
@@ -1,1 +1,0 @@
-export * from './provider-card.component';
