+++ src/app/refill/models/refill-form.ts
@@ -1,0 +1,5 @@
+export interface RefillForm {
+  providerId: string;
+  phoneNumber: string;
+  amount: number;
+}
